0.1.3
Added context menu to the UOBox. Unfortunately, it only works on the current selected UO, not on the one that's being interacted with.

Currently deleting & renaming is supported.

Created Align.m for computing (rigid) transformation between images.

Transitioning between images is now smoother (but more buggy)

Changed the scrolling system and also broke it.