0.1.4
Removed some linking functions from the mlapp file and moved them to GUI or Graphics

Most Graphics functions now take an axisID, this can be used to get the axis object. This way, the program can more easily switch between views. It also enables future support for more views.

The slice per image is now used in most locations instead of app.current_slice. This still needs to implemented in some locations.

Streamlined the scrolling process so that the sliders and the mousewheel now use the same code.
Scrolling can now be done on any view.
The scroll speed is increased based on how quickly the mousewheel is moved.

ctrl+scroll zoom is implemented, but it's still a little buggy.

Min and max scaling is now stored per image.

Added code for copying ROIs between images. There seem to still be some issues with interp3 + it's verrry slow.