roi-4
When editing an ROI, you can now 'cancel' to stop the process. Removed the 'delete ROI' function as it would only cause issues.

undo redo seems stable for polygons, but not always for ellipses and circles. Unsure why.

Still to do:
-contextmenu when right-clicking contour

Will pull this branch after. Not sure when I might work on the contextmenus again.