wc 0.6
Did some code magic to get the ROIs to display correct. The implementation is quite bad and needs to be fixed..

UOs can now be right-clicked to have a context menu appear!

UO info is displayed in the bottom, instead of a stupid floating text box.

ROIprompt is now stored in the background to make launching it quicker.